Fix overflow bugs + add possibility to attack in UI

+++ contracts/src/models/config.cairo
@@ -245,14 +245,14 @@ impl BuildingConfigImpl of BuildingConfigTrait {
 struct TroopConfig {
     #[key]
     config_id: u128,
-    knight_health: u32,
-    paladin_health: u32,
-    crossbowman_health: u32,
-    knight_strength: u32,
-    paladin_strength: u32,
-    crossbowman_strength: u32,
-    advantage_percent: u32,
-    disadvantage_percent: u32,
+    knight_health: u8,
+    paladin_health: u8,
+    crossbowman_health: u8,
+    knight_strength: u8,
+    paladin_strength: u8,
+    crossbowman_strength: u16,
+    advantage_percent: u16,
+    disadvantage_percent: u16,
 }
